requirements.txt
cachetools==5.5.2
Deprecated==1.2.18
docker==7.1.0
python-dotenv==1.0.1
executing==2.2.0
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mlflow==2.20.3
numpy==2.1.3
pandas==2.2.3
PyYAML==6.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
streamlit==1.42.2
transformers
torch
wordcloud
plotly
# onnxruntime
